api/pkg/workspace/{activity,watchers}: move to modules
--- sturdy ---
user_id: cc33c18b-27e5-470d-8589-f9391ca37590
view_id: 2c064d23-2410-4e86-939f-91c56dae2fae
workspace_id: f8fa0c2a-e6ff-4c03-9c23-bc914a63a8f7
