+++ api/pkg/workspace/activity/db/module.go
@@ -1,0 +1,8 @@
+package db
+
+import "getsturdy.com/api/pkg/di"
+
+func Module(c *di.Container) {
+	c.Register(NewActivityReadsRepo)
+	c.Register(NewActivityRepo)
+}
