+++ api/pkg/workspace/activity/sender/module.go
@@ -1,0 +1,7 @@
+package sender
+
+import "getsturdy.com/api/pkg/di"
+
+func Module(c *di.Container) {
+	c.Register(NewActivitySender)
+}
