+++ api/pkg/workspace/watchers/graphql/module.go
@@ -1,8 +1,6 @@
 package graphql
 
-import (
-	"getsturdy.com/api/pkg/di"
-)
+import "getsturdy.com/api/pkg/di"
 
 func Module(c *di.Container) {
 	c.Register(NewRootResolver)
