+++ api/pkg/workspace/watchers/service/module.go
@@ -1,0 +1,7 @@
+package service
+
+import "getsturdy.com/api/pkg/di"
+
+func Module(c *di.Container) {
+	c.Register(New)
+}
